pipeline/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.0.0
pyarrow>=14.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
joblib==1.3.2
apache-airflow==2.7.3
xgboost==2.0.2
scikit-learn==1.3.2
numpy>=1.24.0
python-dotenv==1.0.0
pydantic==2.5.0
